Move fixtures logic to own command/package
